Перемещает логику создания ивента в point-controller.js
